LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_auxclick_is_a_pointerevent-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented


Harness Error (FAIL), message = Unhandled rejection: unimplemented

TIMEOUT auxclick using mouse is a PointerEvent Test timed out
NOTRUN auxclick using pen is a PointerEvent
